REFACTOR :: SecurityConfig
