+++ src/main/java/com/example/demo/global/config/security/SecurityConfig.java
@@ -1,11 +1,13 @@
-package com.example.demo.global.config;
+package com.example.demo.global.config.security;
 
 import com.example.demo.global.auth.CustomUserDetailService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
@@ -26,5 +28,19 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
         return new BCryptPasswordEncoder();
     }
 
-
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+        http
+                .httpBasic().disable()
+                .csrf().disable()
+                .sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS)
+                .and()
+                .authorizeRequests()
+                .antMatchers("/user/**").permitAll()
+                .antMatchers("/auth/logout").authenticated()
+                .antMatchers("/auth/**").permitAll()
+                .antMatchers("/email/delete").authenticated()
+                .antMatchers("/email/**").permitAll()
+                .anyRequest().authenticated();
+    }
 }
