+++ src/main/java/com/example/demo/global/config/security/SecurityUtil.java
@@ -1,0 +1,4 @@
+package com.example.demo.global.config.security;
+
+public class SecurityUtil {
+}
